bug fix in scoreboard (score always == 10) - used doubles instead of longs - and improved eqn

+++ Phase1/GameCentre/app/src/main/java/fall2018/csc2017/GameCentre/BoardManager.java
@@ -18,7 +18,7 @@ class BoardManager implements Serializable {
      */
     private Board board;
     public ScoreBoard scoreBoard;
-    private long score;
+    private double score;
     private User user;
     private SlidingTileSettings slidingTileSettings;
     private Stack<int[]> moves = new Stack<>();
@@ -207,7 +207,7 @@ class BoardManager implements Serializable {
     }
 
 
-    public long getScore() {
+    public double getScore() {
         return this.score;
     }
