+++ Phase1/GameCentre/app/src/main/java/fall2018/csc2017/GameCentre/MovementController.java
@@ -19,7 +19,7 @@ public class MovementController {
         if (boardManager.isValidTap(position)) {
             boardManager.touchMove(position);
             if (boardManager.puzzleSolved()) {
-                long newScore = boardManager.getScore();
+                double newScore = boardManager.getScore();
                 Toast.makeText(context, "YOU WIN!", Toast.LENGTH_SHORT).show();
                 Toast.makeText(context, "you scored=" + newScore, Toast.LENGTH_SHORT).show();
             }
